# frontend/src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Glory RS API",
    "description": "Template API — Rust + Axum + OpenAPI",
    "contact": {
      "name": "1ndoryu"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notifications/{id}/read": {
      "put": {
        "tags": [
          "notifications"
        ],
        "operationId": "mark_read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de notificación",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/timeline": {
      "post": {
        "tags": [
          "timeline"
        ],
        "operationId": "send_timeline_message",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimelineMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineItem"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/timeline/count/{item_type}/{item_id}": {
      "get": {
        "tags": [
          "timeline"
        ],
        "operationId": "count_timeline",
        "parameters": [
          {
            "name": "item_type",
            "in": "path",
            "description": "Tipo de elemento",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "description": "ID legacy del elemento",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineCountResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/timeline/events": {
      "post": {
        "tags": [
          "timeline"
        ],
        "operationId": "create_timeline_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimelineEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/timeline/read": {
      "post": {
        "tags": [
          "timeline"
        ],
        "operationId": "mark_timeline_read",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkTimelineReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineMutationResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/timeline/unread/{item_type}/{item_id}": {
      "get": {
        "tags": [
          "timeline"
        ],
        "operationId": "unread_timeline",
        "parameters": [
          {
            "name": "item_type",
            "in": "path",
            "description": "Tipo de elemento",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "description": "ID legacy del elemento",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineUnreadResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/timeline/{item_type}/{item_id}": {
      "get": {
        "tags": [
          "timeline"
        ],
        "operationId": "list_timeline",
        "parameters": [
          {
            "name": "item_type",
            "in": "path",
            "description": "Tipo de elemento",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "description": "ID legacy del elemento",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateTimelineEventRequest": {
        "type": "object",
        "required": [
          "itemType",
          "itemId",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "itemType": {
            "type": "string"
          },
          "metadata": {
            "nullable": true
          }
        }
      },
      "CreateTimelineMessageRequest": {
        "type": "object",
        "required": [
          "itemType",
          "itemId",
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "itemType": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Estructura de respuesta de error expuesta en la API",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Tipo de error (`not_found`, `unauthorized`, etc.)"
          },
          "message": {
            "type": "string",
            "description": "Mensaje legible para el usuario"
          }
        }
      },
...
      "MarkTimelineReadRequest": {
        "type": "object",
        "required": [
          "itemType",
          "itemId"
        ],
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "itemType": {
            "type": "string"
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "required": [
          "id",
          "notificationType",
          "title",
          "read",
          "createdAt",
          "metadata"
        ],
        "properties": {
          "content": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {},
          "notificationType": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "TimelineCountResponse": {
        "type": "object",
        "required": [
          "total"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TimelineItem": {
        "type": "object",
        "required": [
          "id",
          "itemType",
          "itemId",
          "userId",
          "userName",
          "messageType",
          "content",
          "metadata",
          "createdAt",
          "isOwn"
        ],
        "properties": {
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isOwn": {
            "type": "boolean"
          },
          "itemId": {
            "type": "integer",
            "format": "int64"
          },
          "itemType": {
            "type": "string"
          },
          "messageType": {
            "type": "string"
          },
          "metadata": {},
          "systemAction": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string"
          }
        }
      },
      "TimelineMutationResponse": {
        "type": "object",
        "required": [
          "success",
          "created"
        ],
        "properties": {
          "created": {
            "type": "boolean"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "TimelineResponse": {
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "hasMore"
        ],
        "properties": {
          "hasMore": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "TimelineUnreadResponse": {
        "type": "object",
        "required": [
          "unread"
        ],
        "properties": {
          "unread": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    }
  }
}